References Request handle nested arrays
Validation rules applied on nested arrays are applied globally. I wanted it
to be applied on each deepest array separately for ActorKinships References
(the rule was : kinships.*.references.*).

Let's take an example. An Actor can have several ActorKinships, and each
ActorKinship can have several References. References have to be different for
each ActorKinship, but References can reference several ActorKinships at once.
To ensure References were different, we used to apply a validation rule like
this :

kinships.*.references.* => 'distinct'

But the 'distinct' rule is applied globally, so this array will not pass :

[
  'kinships' => [
    [ 'references' => [1, 2] ],
    [ 'references' => [1] ],
  ]
]

because '1' appears more than once inside the WHOLE array. We wish to have it
applied on each 'references' separately.

I searched how to create custom validation rules in the Laravel docs
(https://laravel.com/docs/6.x/validation#custom-validation-rules).
I found several ways to do it, using either a custom Rule, or using a Service
Provider. But the best solution in my opinion was to loop over every nested
array and to apply the rules on every one of them separatly. This is not the
perfect solution, as we have to loop over the Request fields and add rules
dynamically, but fortunatly, it allows us to use the default 'distinct' rules
on it.