setReferences requires an array instead of a Request object
Requiring a Request object is a strong constraint. It forces to always keep the
root Request, and provide a full path when using a subpart of the request.
Instead, the setReferences method can receive a simpler object like an array,
and we can use the request as an array too.